Allow random sub-sampling of grid square images (#10)
The `SmartEMDiskDataLoader` now allows a number of sub samples and a sub sample size to be specified. If specified the `DataLoader` will produce that many random sub samples of the grid square image with labels calculated as averages over the foil holes contained within the given sub sample. (As this uses the stage coordinates and the coordinates for all AFIS groups are the same this will not be a perfect matching between foil holes and sub samples but should be close enough to provide decent labeling.)