Synchronise the data loaders from disk and from DB so that they have the same behaviour (#11)
The data loaders that get data directly from the PostgreSQL database or from an exported set of images + `yaml` file have diverged and needed to be brought back to behaviour parity. This was best achieved through some relatively small refactoring.

One thing to note is that only projects with the same coordinate calibration are supported within a single data loader at the moment.